Restore ability to remove user installed dictionary files.

+++ src/preferences_dialog.cpp
@@ -1,6 +1,6 @@
 /***********************************************************************
  *
- * Copyright (C) 2008, 2009, 2010, 2011, 2012, 2013, 2014 Graeme Gott <graeme@gottcode.org>
+ * Copyright (C) 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2016 Graeme Gott <graeme@gottcode.org>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -308,6 +308,12 @@ void PreferencesDialog::accept()
 
 	ActionManager::instance()->setShortcuts(m_new_shortcuts);
 
+	// Uninstall languages
+	for (const QString& language : m_uninstalled) {
+		QFile::remove("dict:" + language + ".aff");
+		QFile::remove("dict:" + language + ".dic");
+	}
+
 	// Install languages
 	QString path = DictionaryManager::path() + "/install/";
 	QString new_path = DictionaryManager::installedPath() + "/";
@@ -498,7 +504,7 @@ void PreferencesDialog::addLanguage()
 			QString new_aff_file = dictionary_new_path + language + ".aff";
 			QString new_dic_file = dictionary_new_path + language + ".dic";
 
-			if ((QFile::exists(new_aff_file) || QFile::exists(new_dic_file))) {
+			if (!m_uninstalled.contains(language) && (QFile::exists(new_aff_file) || QFile::exists(new_dic_file))) {
 				if (QMessageBox::question(this, tr("Question"), tr("The dictionary \"%1\" already exists. Do you want to replace it?").arg(name), QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::No) {
 					QFile::remove(aff_file);
 					QFile::remove(dic_file);
@@ -516,6 +522,30 @@ void PreferencesDialog::addLanguage()
 
 	// Close archive
 	zip.close();
+}
+
+//-----------------------------------------------------------------------------
+
+void PreferencesDialog::removeLanguage()
+{
+	int index = m_languages->currentIndex();
+	if (index == -1) {
+		return;
+	}
+	if (QMessageBox::question(this, tr("Question"), tr("Remove current dictionary?"), QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::Yes) {
+		m_uninstalled.append(m_languages->itemData(index).toString());
+		m_languages->removeItem(index);
+	}
+}
+
+//-----------------------------------------------------------------------------
+
+void PreferencesDialog::selectedLanguageChanged(int index)
+{
+	if (index != -1) {
+		QFileInfo info("dict:" + m_languages->itemData(index).toString() + ".dic");
+		m_remove_language_button->setEnabled(info.canonicalFilePath().startsWith(DictionaryManager::installedPath()));
+	}
 }
 
 //-----------------------------------------------------------------------------
@@ -931,10 +961,14 @@ QWidget* PreferencesDialog::initSpellingTab()
 	QGroupBox* languages_group = new QGroupBox(tr("Language"), tab);
 
 	m_languages = new QComboBox(languages_group);
+	connect(m_languages, SIGNAL(currentIndexChanged(int)), this, SLOT(selectedLanguageChanged(int)));
 
 	m_add_language_button = new QPushButton(tr("Add"), languages_group);
 	m_add_language_button->setAutoDefault(false);
 	connect(m_add_language_button, SIGNAL(clicked()), this, SLOT(addLanguage()));
+	m_remove_language_button = new QPushButton(tr("Remove"), languages_group);
+	m_remove_language_button->setAutoDefault(false);
+	connect(m_remove_language_button, SIGNAL(clicked()), this, SLOT(removeLanguage()));
 
 	QStringList languages = DictionaryManager::instance().availableDictionaries();
 	for (const QString& language : languages) {
@@ -946,6 +980,7 @@ QWidget* PreferencesDialog::initSpellingTab()
 	QHBoxLayout* languages_layout = new QHBoxLayout(languages_group);
 	languages_layout->addWidget(m_languages, 1);
 	languages_layout->addWidget(m_add_language_button);
+	languages_layout->addWidget(m_remove_language_button);
 
 	// Read personal dictionary
 	QGroupBox* personal_dictionary_group = new QGroupBox(tr("Personal Dictionary"), tab);
